WEB-16: data-access > 계정 엔티티 id 타입 수정

+++ server/data-access/src/main/kotlin/com/weplay/dataaccess/entity/Account.kt
@@ -27,7 +27,7 @@ import java.time.Instant
 @EntityListeners(AuditingEntityListener::class)
 class Account(
     @Id @GeneratedValue(strategy = GenerationType.AUTO)
-    val id: Long?,
+    val id: Int?,
     val uid: String,
     val status: AccountStatusType,
     val name: String,
